Move Targetable*, Seeds/StoresResource(s) and two RenderBuilding traits to Mods.Common
