+++ OpenRA.Mods.Common/Traits/Render/RenderBuildingSilo.cs
@@ -11,10 +11,9 @@
 using System.Collections.Generic;
 using OpenRA.Graphics;
 using OpenRA.Mods.Common.Graphics;
-using OpenRA.Mods.Common.Traits;
 using OpenRA.Traits;
 
-namespace OpenRA.Mods.RA.Traits
+namespace OpenRA.Mods.Common.Traits
 {
 	class RenderBuildingSiloInfo : RenderBuildingInfo
 	{
