+++ OpenRA.Mods.Common/Traits/StoresResources.cs
@@ -12,7 +12,7 @@ using System.Collections.Generic;
 using System.Linq;
 using OpenRA.Traits;
 
-namespace OpenRA.Mods.RA
+namespace OpenRA.Mods.Common.Traits
 {
 	[Desc("Used for silos.")]
 	class StoresResourcesInfo : ITraitInfo
